Allow types.is_bytes() and types.is_bytearray() to work. Introduce test_types test
