+++ larky/src/main/resources/stdlib/builtins.star
@@ -4,16 +4,15 @@ load("@stdlib/codecs", "codecs")
 
 
 def _bytes(s, encoding='utf-8', errors='strict'):
-    """
-    While bytes literals and representations are based on ASCII text, bytes
+    """While bytes literals and representations are based on ASCII text, bytes
     objects actually behave like immutable sequences of integers, with each
-    value in the sequence restricted such that 0 <= x < 256 (attempts to violate
-    this restriction will trigger ValueError). This is done deliberately to
-    emphasise that while many binary formats include ASCII based elements and
-    can be usefully manipulated with some text-oriented algorithms, this is not
-    generally the case for arbitrary binary data (blindly applying text
-    processing algorithms to binary data formats that are not ASCII compatible
-    will usually lead to data corruption).
+    value in the sequence restricted such that 0 <= x < 256 (attempts to
+    violate this restriction will trigger ValueError). This is done
+    deliberately to emphasise that while many binary formats include ASCII
+    based elements and can be usefully manipulated with some text-oriented
+    algorithms, this is not generally the case for arbitrary binary data
+    (blindly applying text processing algorithms to binary data formats that
+    are not ASCII compatible will usually lead to data corruption).
 
     In addition to the literal forms, bytes objects can be created in a number
     of other ways:
@@ -45,7 +44,33 @@ def _bytes(s, encoding='utf-8', errors='strict'):
     #     fail("TypeError: cannot convert '%s' object to bytes" % type(s))
 
 
+def _bytearray(source, encoding='utf-8', errors='strict'):
+    """Return a new array of bytes. The bytearray class is a mutable sequence
+    of integers in the range 0 <= x < 256. It has most of the usual methods of
+    mutable sequences, described in Mutable Sequence Types, as well as most
+    methods that the bytes type has, see Bytes and Bytearray Operations.
+
+    The optional source parameter can be used to initialize the array in a few
+    different ways:
+
+        - If it is a string, you must also give the encoding
+        (and optionally, errors) parameters; bytearray() then converts the
+        string to bytes using str.encode().
+        - If it is an integer, the array will have that size and will be
+        initialized with null bytes.
+        - If it is an object conforming to the buffer interface, a read-only
+        buffer of the object will be used to initialize the bytes array.
+        - If it is an iterable, it must be an iterable of integers in the
+        range 0 <= x < 256, which are used as the initial contents of the
+        array.
+
+    Without an argument, an array of size 0 is created.
+    """
+    return larky.bytearray(source, encoding, errors)
+
+
 builtins = larky.struct(
     bytes=_bytes,
-    b=_bytes
+    b=_bytes,
+    bytearray=_bytearray,
 )
